UnicodeInputHelper/requirements.txt
keyboard
clipboard
Pillow
pystray